requirements.txt
discord.py
fastapi
uvicorn
google-generativeai
PyYAML